+++ src/routes/RequestsRoute.js
@@ -70,6 +70,7 @@ import {
   ITEM_QUERIES,
   PRINT_DETAILS_COLUMNS,
   RESOURCE_TYPES,
+  requestFilterTypes,
 } from '../constants';
 import {
   buildUrl,
@@ -106,6 +107,8 @@ import {
   getFormattedYears,
   getStatusQuery,
   getFullNameForCsvRecords,
+  getPrintStatusFilteredData,
+  filterRecordsByPrintStatus,
 } from './utils';
 import SinglePrintButtonForPickSlip from '../components/SinglePrintButtonForPickSlip';
 
@@ -613,8 +616,6 @@ class RequestsRoute extends React.Component {
       createTitleLevelRequestsByDefault = false,
     } = getTlrSettings(props.resources.configs.records[0]?.value);
 
-    const isViewPrintDetailsEnabled = get(props.resources, viewPrintDetailsPath) === 'true';
-
     this.okapiUrl = props.stripes.okapi.url;
 
     this.httpHeadersOptions = {
@@ -648,7 +649,8 @@ class RequestsRoute extends React.Component {
       selectedRows: {},
       titleLevelRequestsFeatureEnabled,
       createTitleLevelRequestsByDefault,
-      isViewPrintDetailsEnabled,
+      isViewPrintDetailsEnabled: false,
+      selectedPrintStatusFilters: [],
     };
 
     this.pickSlipsPrintContentRef = React.createRef();
@@ -657,13 +659,21 @@ class RequestsRoute extends React.Component {
     this.printSelectedContentRef = React.createRef();
   }
 
-  static getDerivedStateFromProps(props) {
+  static getDerivedStateFromProps(props, state) {
     const layer = (props.resources.query || {}).layer;
+    const newState = {};
+    const currViewPrintDetailsSettings = get(props.resources, viewPrintDetailsPath) === 'true';
+
     if (!layer) {
-      return { dupRequest: null };
+      newState.dupRequest = null;
     }
 
-    return null;
+    if (currViewPrintDetailsSettings !== state.isViewPrintDetailsEnabled) {
+      // Update the `isViewPrintDetailsEnabled` state based on user navigation back to Request App.
+      newState.isViewPrintDetailsEnabled = currViewPrintDetailsSettings;
+    }
+
+    return Object.keys(newState).length ? newState : null;
   }
 
   componentDidMount() {
@@ -677,7 +687,11 @@ class RequestsRoute extends React.Component {
   }
 
   componentDidUpdate(prevProps) {
-    const { submitting } = this.state;
+    const {
+      submitting,
+      isViewPrintDetailsEnabled,
+      selectedPrintStatusFilters,
+    } = this.state;
     const { stripes } = this.props;
     const patronBlocks = get(this.props.resources, ['patronBlocks', 'records'], []);
     const prevBlocks = get(prevProps.resources, ['patronBlocks', 'records'], []);
@@ -686,19 +700,8 @@ class RequestsRoute extends React.Component {
     const { id: currentServicePointId } = this.getCurrentServicePointInfo();
     const prevStateServicePointId = get(prevProps.resources.currentServicePoint, 'id');
     const { configs: prevConfigs } = prevProps.resources;
-    const { configs } = this.props.resources;
+    const { configs, query: { filters } } = this.props.resources;
     const instanceId = parse(this.props.location?.search)?.instanceId;
-    const currPrintDetailsSettings = get(this.props.resources, viewPrintDetailsPath);
-    const prevPrintDetailsSettings = get(prevProps.resources, viewPrintDetailsPath);
-
-    if (currPrintDetailsSettings !== prevPrintDetailsSettings) {
-      const isViewPrintDetailsEnabled = currPrintDetailsSettings === 'true';
-      this.setState({ isViewPrintDetailsEnabled });
-
-      if (!isViewPrintDetailsEnabled) {
-        this.columnHeadersMap = getFilteredColumnHeadersMap(this.columnHeadersMap);
-      }
-    }
 
     if (prevExpired.length > 0 && expired.length === 0) {
       // eslint-disable-next-line react/no-did-update-set-state
@@ -739,6 +742,16 @@ class RequestsRoute extends React.Component {
       this.onSearchComplete(this.props.resources.records);
     }
 
+    if (isViewPrintDetailsEnabled !== prevState.isViewPrintDetailsEnabled && !isViewPrintDetailsEnabled) {
+      this.columnHeadersMap = getFilteredColumnHeadersMap(this.columnHeadersMap);
+    }
+
+    if (filters?.includes(requestFilterTypes.PRINT_STATUS)) {
+      const printStatusFilterInQuery = this.getActiveFilters()[requestFilterTypes.PRINT_STATUS];
+
+      this.updateSelectedPrintStatusFilters(isViewPrintDetailsEnabled, selectedPrintStatusFilters, printStatusFilterInQuery);
+    }
+
     if (stripes?.user?.user?.tenants && stripes.user.user !== prevProps.stripes?.user?.user) {
       this.getEcsTlrSettings();
     }
@@ -772,6 +785,24 @@ class RequestsRoute extends React.Component {
             isEcsTlrSettingEnabled: false,
           });
         });
+    }
+  }
+
+  updateSelectedPrintStatusFilters(isViewPrintDetailsEnabled, selectedPrintStatusFilters, printStatusFilterInQuery) {
+    /**
+     * Updates the `selectedPrintStatusFilters` state based on pre selected filters when user navigates back to Request App.
+     *
+     * The function performs the following actions:
+     * 1. If `isViewPrintDetailsEnabled` is true and if `Print Status` filters contains exactly one filter:
+     *    - it updates state to set `selectedPrintStatusFilters` to this one `Print Status` filter.
+     *
+     * 2. If `isViewPrintDetailsEnabled` is false and `filters` in query includes 'PRINT STATUS' filter:
+     *    - it clears the 'PRINT STATUS' filter from query by invoking `handleFilterChange`.
+     */
+    if (isViewPrintDetailsEnabled && selectedPrintStatusFilters.length === 0 && printStatusFilterInQuery?.length === 1) {
+      this.setState({ selectedPrintStatusFilters: [printStatusFilterInQuery[0]] });
+    } else if (!isViewPrintDetailsEnabled && printStatusFilterInQuery?.length) {
+      this.handleFilterChange({ name: requestFilterTypes.PRINT_STATUS, values: [] });
     }
   }
 
@@ -849,6 +880,11 @@ class RequestsRoute extends React.Component {
   // Export function for the CSV search report action
   async exportData() {
     this.setState({ csvReportPending: true });
+    const { isViewPrintDetailsEnabled, selectedPrintStatusFilters } = this.state;
+
+    const activeFilters = this.getActiveFilters();
+    const activeFilterKeys = Object.keys(activeFilters);
+    const isOnlyPrintStatusFilterSelected = activeFilterKeys.length === 1 && activeFilterKeys[0] === requestFilterTypes.PRINT_STATUS;
 
     // Build a custom query for the CSV record export, which has to include
     // all search and filter parameters
@@ -856,17 +892,21 @@ class RequestsRoute extends React.Component {
     let queryString;
 
     const queryTerm = this.props.resources?.query?.query;
-    const filterQuery = filters2cql(RequestsFiltersConfig, deparseFilters(this.getActiveFilters()));
+    const filterQuery = filters2cql(RequestsFiltersConfig, deparseFilters(activeFilters));
 
     if (queryTerm) {
       queryString = `(requesterId=="${queryTerm}" or requester.barcode="${queryTerm}*" or item.title="${queryTerm}*" or item.barcode=="${queryTerm}*" or itemId=="${queryTerm}")`;
       queryClauses.push(queryString);
     }
     if (filterQuery) queryClauses.push(filterQuery);
+    if (isOnlyPrintStatusFilterSelected) queryClauses.push('cql.allRecords=1');
 
     queryString = queryClauses.join(' and ');
     const records = await this.fetchReportData(this.props.mutator.reportRecords, queryString);
-    const recordsToCSV = this.buildRecords(records);
+
+    const printStatusFilteredRecords = isViewPrintDetailsEnabled && selectedPrintStatusFilters.length === 1 &&
+      filterRecordsByPrintStatus(records, selectedPrintStatusFilters);
+    const recordsToCSV = this.buildRecords(printStatusFilteredRecords || records);
 
     exportCsv(recordsToCSV, {
       onlyFields: this.columnHeadersMap,
@@ -1239,6 +1279,10 @@ class RequestsRoute extends React.Component {
   }
 
   handleFilterChange = ({ name, values }) => {
+    if (name === requestFilterTypes.PRINT_STATUS) {
+      this.setState({ selectedPrintStatusFilters: values });
+    }
+
     const { mutator } = this.props;
     const newFilters = {
       ...this.getActiveFilters(),
@@ -1279,7 +1323,7 @@ class RequestsRoute extends React.Component {
 
   renderFilters = (onChange) => {
     const { resources } = this.props;
-    const { titleLevelRequestsFeatureEnabled } = this.state;
+    const { titleLevelRequestsFeatureEnabled, isViewPrintDetailsEnabled } = this.state;
 
     return (
       <RequestsFilters
@@ -1288,6 +1332,7 @@ class RequestsRoute extends React.Component {
         onChange={onChange}
         onClear={(name) => onChange({ name, values: [] })}
         titleLevelRequestsFeatureEnabled={titleLevelRequestsFeatureEnabled}
+        isViewPrintDetailsEnabled={isViewPrintDetailsEnabled}
       />
     );
   };
@@ -1380,6 +1425,7 @@ class RequestsRoute extends React.Component {
       isViewPrintDetailsEnabled,
       isEcsTlrSettingReceived,
       isEcsTlrSettingEnabled,
+      selectedPrintStatusFilters,
     } = this.state;
     const isPrintHoldRequestsEnabled = getPrintHoldRequestsEnabled(resources.printHoldRequests);
     const { name: servicePointName } = this.getCurrentServicePointInfo();
@@ -1425,6 +1471,7 @@ class RequestsRoute extends React.Component {
 
     const isPickSlipsArePending = resources?.pickSlips?.isPending;
     const isSearchSlipsArePending = resources?.searchSlips?.isPending;
+    const isRequestsRecordsLoaded = resources.records.hasLoaded;
     const requestsEmpty = isEmpty(requests);
     const isPickSlipsEmpty = isEmpty(pickSlips);
     const isSearchSlipsEmpty = isEmpty(searchSlips);
@@ -1433,6 +1480,14 @@ class RequestsRoute extends React.Component {
     const pickSlipsData = convertToSlipData(pickSlips, intl, timezone, locale, SLIPS_TYPE.PICK_SLIP, user);
     const searchSlipsData = convertToSlipData(searchSlips, intl, timezone, locale, SLIPS_TYPE.SEARCH_SLIP_HOLD_REQUESTS);
     let multiSelectPickSlipData = getSelectedSlipDataMulti(pickSlipsData, selectedRows);
+    /**
+     * For 'displayPrintStatusFilteredData' to be true the length of 'selectedPrintStatusFilters' must be 1.
+     * This is because we only filter data when exactly one PrintStatus filter is selected ([Printed] or [Not Printed]).
+     * If the filter array is empty or contains both filters ([] or [Printed, Not Printed]),
+     * no filtering is needed as the data should be used directly from the query response.
+     */
+    const displayPrintStatusFilteredData = isViewPrintDetailsEnabled &&
+    isRequestsRecordsLoaded && selectedPrintStatusFilters.length === 1;
 
     const resultsFormatter = getListFormatter(
       {
@@ -1654,7 +1709,8 @@ class RequestsRoute extends React.Component {
               customPaneSub={this.renderPaneSub()}
               onCreate={this.create}
               onCloseNewRecord={this.handleCloseNewRecord}
-              parentResources={resources}
+              parentResources={displayPrintStatusFilteredData ?
+                getPrintStatusFilteredData(resources, selectedPrintStatusFilters) : resources}
               parentMutator={mutator}
               detailProps={{
                 onChangePatron: this.onChangePatron,